@KristofferC 2019-06-11
This doesn't seem safe in general? What if `x` goes out of scope before the `unsafe_wrap` Array.