@SimonDanisch 2019-06-11
Guess it's unsafe indeed :P 
I was assuming, it's only used just before sending it to write, but it's a bit hard to tell where things go.
Maybe it'd be better to implement a write that works with an reinterpret array efficiently.