@KristofferC 2019-06-11
>  Maybe it'd be better to implement a write that works with an reinterpret array efficiently.

Yeah, probably. https://github.com/MikeInnes/BSON.jl/pull/46 is at least a lot better than what we have now.