@SimonDanisch 2019-06-11
Yup, I was thinking to just submit something like #46, but then I got greedy :D