@SimonDanisch 2019-06-11
I'd say just merge #46 and leave this unmerged until we find something safer ;)