@KristofferC 2019-06-11
Interestingly, just defining

```jl
function reinterpret_(::Type{T}, x) where T
  reinterpret(T, x)
end
```

I can load VGG19 pretty damn fast

```jl
julia> @time Metalhead.weights("vgg19.bson")
  0.269944 seconds (1.97 k allocations: 548.196 MiB, 37.98% gc time)
Dict{Symbol,Any} with 38 entries:
```

but it crashes when testing BSON itself.