@richiejp 2019-06-24
@KristofferC probably because reinterpret returns `ReinterpretArray` which wraps the original array and lazily casts its contents when accessed.